ci-operator: handle labels in a DRY manner, better rehearsal contexts
